pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "diffusion-tinker"
dynamic = ["version"]
description = "RL-based post-training for diffusion models"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "diffusers>=0.31.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "peft>=0.12.0",
    "Pillow",
    "tqdm",
    "huggingface-hub>=0.20.0",
    "safetensors>=0.4.0",
]

[project.optional-dependencies]
rewards = ["open-clip-torch"]
wandb = ["wandb"]
dev = [
    "diffusion-tinker[rewards,wandb]",
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "diffusion_tinker._version.__version__"}

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
